타인 열람시 삭제 기능 비활성화 수정

+++ src/components/shared/article/index.tsx
@@ -14,6 +14,7 @@ type ArticleTypes = Omit<ArticleType, 'user'>;
 interface ArticleProps {
   threedot?: ReactNode;
   statusFlag?: string;
+  isOwner?: boolean;
   article: ArticleTypes;
 }
 
@@ -40,11 +41,15 @@ const formatContent = (content: string, maxLines: number = Infinity): ReactNode 
   );
 };
 
-const Article = ({ article, statusFlag }: ArticleProps) => {
+const Article = ({ article, statusFlag, isOwner }: ArticleProps) => {
   const navigate = useNavigate();
   const { open } = useModalContext();
 
-  const { likeCount, likeType, toggleLike } = useLike(article.postId, article.likeCount, article.likeType);
+  const { likeCount, likeType, toggleLike } = useLike(
+    article.postId,
+    article.likeCount,
+    article.likeType,
+  );
 
   const iconValue: IconValues = likeType ? 'yeslike' : 'nolike';
 
@@ -60,27 +65,41 @@ const Article = ({ article, statusFlag }: ArticleProps) => {
   const onHandleSetting = (e: React.MouseEvent) => {
     e.stopPropagation();
 
-    open({
-      title: '설정',
-      type: 'setting',
-      buttonLabel: '삭제',
-      onButtonClick: async () => {
-        try {
-          await deleteArticle(article.postId);
-        } catch (error) {
-          console.error(error);
-        }
-      },
-    });
-  }
-  
+    if (isOwner) {
+      open({
+        title: '설정',
+        type: 'setting',
+        buttonLabel: '삭제',
+        flag: true,
+        onButtonClick: async () => {
+          
+        },
+      });
+    } else {
+      open({
+        title: '설정',
+        type: 'setting',
+        flag: false,
+        onButtonClick: async () => {
+          try {
+            await deleteArticle(article.postId);
+          } catch (error) {
+            console.error(error);
+          }
+        },
+      });
+    }
+    
+  };
 
   return (
     <S.ArticleContainer onClick={toHandleDetail}>
       <S.ArTopBox>
-        <Text typography='t1' bold='regular' color='black'>{article.title}</Text>
+        <Text typography="t1" bold="regular" color="black">
+          {article.title}
+        </Text>
         {statusFlag === 'open' && (
-          <Icon value='threedot' color='black' onClick={(e) => onHandleSetting(e)}/>
+          <Icon value="threedot" color="black" onClick={(e) => onHandleSetting(e)} />
         )}
       </S.ArTopBox>
       <S.ArContentBlock>{contentNode}</S.ArContentBlock>
@@ -93,8 +112,8 @@ const Article = ({ article, statusFlag }: ArticleProps) => {
             toggleLike();
           }}
         />
-        <InfoStatus value="comment" count={article.commentCount} $active={false}/>
-        <InfoStatus value="check" count={article.codeCount} $active={false}/>
+        <InfoStatus value="comment" count={article.commentCount} $active={false} />
+        <InfoStatus value="check" count={article.codeCount} $active={false} />
       </S.ArDataBlock>
     </S.ArticleContainer>
   );
